fix history classes in backend

+++ Code/fitervari_backend/src/main/java/at/htl_leonding/model/Set.kt
@@ -11,7 +11,7 @@ class Set(
         var weight: Double,
         var time: Double,
         @Column(name = "set_number")
-        var setNumber: Double,
+        var setNumber: Int,
         var type: String
 ) : PanacheEntity() {
     fun copyValues(other: Set) {
